Save Level Bug Fix

+++ Assets/Scripts/Runtime/Keys/SaveGameDataParams.cs
@@ -6,7 +6,7 @@ namespace Runtime.Keys
     public struct SaveGameDataParams
     {
         public int Level;
-        public float Money;
+        public int Money;
         public int IncomeLevel;
         public int StackLevel;
     }
